rename circuit-value.ts to struct.ts
