+++ src/lib/foreign-field.ts
@@ -13,7 +13,7 @@ import { Gadgets } from './gadgets/gadgets.js';
 import { ForeignField as FF } from './gadgets/foreign-field.js';
 import { assert } from './gadgets/common.js';
 import { l3, l } from './gadgets/range-check.js';
-import { ProvablePureExtended } from './provable-types/circuit-value.js';
+import { ProvablePureExtended } from './provable-types/struct.js';
 
 // external API
 export { createForeignField };
