+++ src/lib/gadgets/arithmetic.ts
@@ -1,4 +1,4 @@
-import { provableTuple } from '../provable-types/circuit-value.js';
+import { provableTuple } from '../provable-types/struct.js';
 import { Field } from '../core.js';
 import { assert } from '../errors.js';
 import { Provable } from '../provable.js';
