+++ src/lib/hash.ts
@@ -2,7 +2,7 @@ import {
   HashInput,
   ProvableExtended,
   Struct,
-} from './provable-types/circuit-value.js';
+} from './provable-types/struct.js';
 import { Snarky } from '../snarky.js';
 import { Field } from './core.js';
 import { createHashHelpers } from './hash-generic.js';
