+++ src/lib/merkle-map.ts
@@ -1,4 +1,4 @@
-import { arrayProp, CircuitValue } from './provable-types/circuit-value.js';
+import { arrayProp, CircuitValue } from './provable-types/struct.js';
 import { Field, Bool } from './core.js';
 import { Poseidon } from './hash.js';
 import { MerkleTree, MerkleWitness } from './merkle-tree.js';
