+++ src/lib/provable-types/merkle-list.ts
@@ -1,6 +1,6 @@
 import { Bool, Field } from '../core.js';
 import { Provable } from '../provable.js';
-import { Struct, Unconstrained } from './circuit-value.js';
+import { Struct, Unconstrained } from './struct.js';
 import { assert } from '../gadgets/common.js';
 import { provableFromClass } from '../../bindings/lib/provable-snarky.js';
 import { Poseidon, packToFields, ProvableHashable } from '../hash.js';
